+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added Markdown export
 - Added METAR export
 - Fixed cloud bug which brakes the `main.mcf`
+- Markdown now changes to minute output if flight plan is less than an hour
 
 ## [1.7.0] - 2026-06-01